moved view count calculation into door43_tools
