+++ libraries/lambda_handlers/page_view_count_handler.py
@@ -1,8 +1,8 @@
 from __future__ import unicode_literals, print_function
 import json
 import urllib
+from libraries.door43_tools.page_metrics import PageMetrics
 from libraries.lambda_handlers.handler import Handler
-from libraries.manifest.page_metrics import PageMetrics
 
 
 class PageViewCountHandler(Handler):
