+++ tests/door43_tools_tests/test_viewCountHandler.py
@@ -1,10 +1,9 @@
 from __future__ import absolute_import, unicode_literals, print_function
 import json
 import mock
-from unittest import TestCase
 from decimal import Decimal
-from libraries.lambda_handlers.view_count_handler import ViewCountHandler
-from libraries.manifest.page_metrics import PageMetrics
+from unittest import TestCase
+from libraries.lambda_handlers.page_view_count_handler import PageViewCountHandler
 
 
 class ViewCountHandlerTest(TestCase):
@@ -15,7 +14,7 @@ class ViewCountHandlerTest(TestCase):
         self.return_view_count = 0
         self.error_response = None
 
-    @mock.patch('libraries.manifest.page_metrics.PageMetrics.get_view_count')
+    @mock.patch('libraries.door43_tools.page_metrics.PageMetrics.get_view_count')
     def test_handle(self, mock_get_view_count):
         # given
         mock_get_view_count.side_effect = self.mock_get_view_count
@@ -24,7 +23,7 @@ class ViewCountHandlerTest(TestCase):
         callback = 'cb'
         return_view_count = 1234
         event = self.initialize_test(callback, path, increment, return_view_count)
-        handler = ViewCountHandler()
+        handler = PageViewCountHandler()
 
         # when
         response = handler.handle(event, None)
@@ -32,7 +31,7 @@ class ViewCountHandlerTest(TestCase):
         # then
         self.validate_results(response)
 
-    @mock.patch('libraries.manifest.page_metrics.PageMetrics.get_view_count')
+    @mock.patch('libraries.door43_tools.page_metrics.PageMetrics.get_view_count')
     def test_handleNoIncrement(self, mock_get_view_count):
         # given
         mock_get_view_count.side_effect = self.mock_get_view_count
@@ -41,7 +40,7 @@ class ViewCountHandlerTest(TestCase):
         callback = 'cb'
         return_view_count = 1234
         event = self.initialize_test(callback, path, increment, return_view_count)
-        handler = ViewCountHandler()
+        handler = PageViewCountHandler()
 
         # when
         response = handler.handle(event, None)
@@ -49,7 +48,7 @@ class ViewCountHandlerTest(TestCase):
         # then
         self.validate_results(response)
 
-    @mock.patch('libraries.manifest.page_metrics.PageMetrics.get_view_count')
+    @mock.patch('libraries.door43_tools.page_metrics.PageMetrics.get_view_count')
     def test_handleIncrement(self, mock_get_view_count):
         # given
         mock_get_view_count.side_effect = self.mock_get_view_count
@@ -58,7 +57,7 @@ class ViewCountHandlerTest(TestCase):
         callback = 'cb'
         return_view_count = 1234
         event = self.initialize_test(callback, path, increment, return_view_count)
-        handler = ViewCountHandler()
+        handler = PageViewCountHandler()
         self.expected_view_count = return_view_count + 1  # expect incremented count
 
         # when
